very primitive interface; some compiler warning fixes

+++ include/list.h
@@ -6,7 +6,7 @@
 typedef struct archive_list_node {
     struct archive_list_node *next;
     const char *file_name;
-    unsigned int offset;
+    size_t offset;
     size_t len;
 } archive_list_node_t;
 
@@ -17,8 +17,8 @@ typedef struct {
 } archive_list_t;
 
 archive_list_t *archive_list_init();
-int archive_list_add(archive_list_t *list, const char *file_name,
-                     unsigned int offset, size_t len);
+int archive_list_add(archive_list_t *list, const char *file_name, size_t offset,
+                     size_t len);
 archive_list_node_t *archive_list_get(archive_list_t *list, unsigned int index);
 void archive_list_destroy(archive_list_t *list);
